+++ analysis/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased] (Added 🚀 | Changed | Removed  | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
 
+### Added 🚀
+
+- Add DependaChartaImporter to import code-level dependency data from `.dc.json` files as edge metrics
+
 ## [1.140.0] - 2025-12-05
 
 ### Fixed 🐞